+++ Application/Categories/Queries/GetCategoryDetail/GetCategoryDetailQuery.cs
@@ -1,0 +1,10 @@
+﻿using MediatR;
+using System.Runtime.InteropServices;
+
+namespace SimpleCMS.Application.Categories.Queries.GetCategoryDetail
+{
+    public class GetCategoryDetailQuery : IRequest<CategoryDetailVM>
+    {
+        public int Id { get; set; }
+    }
+}