+++ Application/Common/Mappings/IMapFrom.cs
@@ -9,6 +9,6 @@ namespace SimpleCMS.Application.Common.Mappings
 {
     public interface IMapFrom<T>
     {
-        void Mapping(Profile profile) => profile.CreateMap(typeof(T), GetType());
+        void Mapping(Profile profile);
     }
 }