+++ Application/Contents/Commands/DeleteContent/DeleteContentCommandValidator.cs
@@ -1,0 +1,12 @@
+﻿using FluentValidation;
+
+namespace SimpleCMS.Application.Contents.Commands.DeleteContent
+{
+    public class DeleteContentCommandValidator : AbstractValidator<DeleteContentCommand>
+    {
+        public DeleteContentCommandValidator()
+        {
+            RuleFor(v => v.Id).NotEmpty();
+        }
+    }
+}