+++ Application/Contents/Commands/UpsertContent/UpsertContentCommandValidator.cs
@@ -1,0 +1,13 @@
+﻿using FluentValidation;
+
+namespace SimpleCMS.Application.Contents.Commands.UpsertContent
+{
+    public class UpsertContentCommandValidator : AbstractValidator<UpsertContentCommand>
+    {
+        public UpsertContentCommandValidator()
+        {
+            RuleFor(v => v.TopicId).NotEmpty();
+            RuleFor(v => v.Name).NotEmpty();
+        }
+    }
+}