+++ Application/Contents/Queries/GetContentsList/ContentDTO.cs
@@ -2,16 +2,19 @@
 using SimpleCMS.Application.Common.Mappings;
 using SimpleCMS.Domain.Entities;
 
-namespace SimpleCMS.Application.Categories.Queries.GetCategoriesList
+namespace SimpleCMS.Application.Contents.Queries.GetContentsList
 {
-    public class CategoryDTO : IMapFrom<Category>
+    public class ContentDTO : IMapFrom<Content>
     {
         public int Id { get; set; }
+
+        public int TopicId { get; set; }
+
         public string Name { get; set; }
         public string Description { get; set; }
 
         public void Mapping(Profile profile) =>
-            profile.CreateMap<Category, CategoryDTO>()
-            .ForMember(d => d.Id, opt => opt.MapFrom(e => e.CategoryId));
+            profile.CreateMap<Content, ContentDTO>()
+            .ForMember(d => d.Id, opt => opt.MapFrom(e => e.ContentId));
     }
 }