+++ Application/Topics/Queries/GetTopicDetail/TopicContentDTO.cs
@@ -2,16 +2,17 @@
 using SimpleCMS.Application.Common.Mappings;
 using SimpleCMS.Domain.Entities;
 
-namespace SimpleCMS.Application.Categories.Queries.GetCategoriesList
+namespace SimpleCMS.Application.Topics.Queries.GetTopicDetail
 {
-    public class CategoryDTO : IMapFrom<Category>
+    public class TopicContentDTO : IMapFrom<Content>
     {
         public int Id { get; set; }
+
         public string Name { get; set; }
         public string Description { get; set; }
 
         public void Mapping(Profile profile) =>
-            profile.CreateMap<Category, CategoryDTO>()
-            .ForMember(d => d.Id, opt => opt.MapFrom(e => e.CategoryId));
+            profile.CreateMap<Content, TopicContentDTO>()
+            .ForMember(d => d.Id, opt => opt.MapFrom(e => e.ContentId));
     }
 }